Change SConstruct files. Fix shameful warnings and bugs.

+++ lib/period.cpp
@@ -12,7 +12,7 @@ namespace MTToolBox {
     /**
      * メルセンヌ指数 p のリスト
      */
-    static const uint32_t mersenne_exponent[] =
+    static const int32_t mersenne_exponent[] =
     {2, 3, 5, 7, 13, 17, 19, 31, 61, 89, 107, 127, 521, 607, 1279,
      2203, 2281, 3217, 4253, 4423, 9689, 9941, 11213, 19937, 21701,
      23209, 44497, 86243, 110503, 132049, 216091, 756839, 859433,
@@ -69,7 +69,7 @@ namespace MTToolBox {
     bool
     isMexp(uint32_t degree) {
         for (int i = 0; mersenne_exponent[i] > 0; i++) {
-            if (degree == mersenne_exponent[i]) {
+            if ((int32_t)degree == mersenne_exponent[i]) {
                 return true;
             }
         }
